fix: prevent DB deadlock on bid save
Remove ON CONFLICT, which requires a lock and can cause data loss if a deadlock occurs.
Add a retry mechanism to further reduce the chance of errors leading to lost data.

Remove the uniqueness constraint from block submissions to speed up inserts and reduce lock contention.
This removes built-in deduplication, but we already perform a dedupe when a bid is received.
Some duplicates can still slip through, so add an additional dedupe step in the data API before returning results to the user.

Remove get_header_trace and submission_trace.
get_header times can be flattened into the get_header table, and submission_trace is obsolete since its recorded times have been zero for a while.